ETL_Scraping_DataLakehouse/requirements.txt
# Core Data Engineering
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.2
duckdb==0.9.2

# Web Scraping
selenium==4.15.2
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
fake-useragent==1.4.0
undetected-chromedriver==3.5.4

# Data Lakehouse & Storage
delta-spark==3.0.0
minio==7.2.0
boto3==1.34.0
s3fs==2023.12.0

# DBT & Data Transformation
dbt-core==1.7.3
dbt-duckdb==1.7.3
dbt-utils==1.1.1

# Data Quality & Testing
great-expectations==0.18.19
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Orchestration
apache-airflow==2.7.3
prefect==2.14.8
celery==5.3.4

# API & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Monitoring & Logging
prometheus-client==0.19.0
grafana-api==1.0.3
structlog==23.2.0
colorama==0.4.6

# Configuration & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
croniter==2.0.1

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Jupyter (for development)
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6 